All I did while resting on the sea =)

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -2,20 +2,23 @@
 namespace App\Http\Controllers\Admin;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Controllers\Controller;
-use App\Repositories\Admin\CategoryRepository as Categories;
+use App\Repositories\CategoryRepository as Categories;
 
 class CategoryController extends Controller
 {
 	/**
-	 * Dependency, which will be inject in constructor
+	 * Dependency, which will be inject in constructor.
+	 * 
 	 * @var Categories
 	 */
     protected $categories;
 
-
 	/**
 	 * CategoryController constructor.
+	 * 
 	 * @param Categories $categories
 	 */
 	public function __construct(Categories $categories)
@@ -23,34 +26,38 @@ class CategoryController extends Controller
 		$this->categories = $categories;
 	}
 
-
 	/**
 	 * Get all categories from repository and show them.
 	 * 
 	 * @param Request $request
 	 * @return \Illuminate\Http\Response
 	 * */
-	public function getAll(Request $request)
+	public function allCategories(Request $request)
 	{
 		if ($request->ajax()) {
 			$categories = $this->categories->allCategoriesForAdmin($request->ordered_column, $request->direction);
+//			$categories->pagination_menu = $categories->render();
 			return response()->view('admin.category.table', array('categories' => $categories));
 		}
 
 		$categories = $this->categories->allCategoriesForAdmin();
+//		$categories->pagination_menu = $categories->render();
 		return response()->view('admin.category.all', array('categories' => $categories));
 	}
-
-
+	
 	/**
 	 * Get the form for creating a new category.
+	 * 
 	 * @return \Illuminate\Http\Response
 	 */
-	public function getCreateView()
+	public function createCategoryView()
 	{
+		if(Gate::denies('createCategory', Auth::user())) {
+			return redirect()->back()->with('error', 'no permission');
+		}
+
 		return response()->view('admin.category.create');
 	}
-
 	
 	/**
 	 * Store the new category and redirect 
@@ -59,7 +66,7 @@ class CategoryController extends Controller
 	 * @param Request $request
 	 * @return \Illuminate\Http\RedirectResponse
 	 */
-	public function postCreate(Request $request)
+	public function createCategory(Request $request)
 	{
 		$result = $this->categories->store($request);
 
@@ -72,20 +79,22 @@ class CategoryController extends Controller
 			->with('info', 'The category was created successfully');
 	}
 
-
 	/**
 	 * Get the category by id and show the form for editing it.
 	 *
 	 * @param $id
 	 * @return \Illuminate\Http\Response
 	 */
-	public function getUpdateView($id)
+	public function updateCategoryView($id)
 	{
+		if(Gate::denies('updateCategory', Auth::user())) {
+			return redirect()->back()->with('error', 'no permission');
+		}
+
 		$category = $this->categories->findById($id);
 		return response()->view('admin.category.update', array('category' => $category));
 	}
-
-
+	
 	/**
 	 * Update the category and redirect to 
 	 * the page with all categories with notification of result.
@@ -93,7 +102,7 @@ class CategoryController extends Controller
 	 * @param Request $request
 	 * @return \Illuminate\Http\RedirectResponse
 	 */
-	public function postUpdate(Request $request)
+	public function updateCategory(Request $request)
 	{
 		$result = $this->categories->update($request);
 		if ($result == true) {
@@ -102,8 +111,7 @@ class CategoryController extends Controller
 		
 		return redirect()->route('adminCategories');
 	}
-
-
+	
 	/**
 	 * Delete the category and redirect to
 	 * the page with all categories with notification of result.
@@ -111,11 +119,47 @@ class CategoryController extends Controller
 	 * @param Request $request
 	 * @return \Illuminate\Http\RedirectResponse
 	 */
-	public function postDelete(Request $request)
+	public function deleteCategory(Request $request)
 	{
+		if (!$request->ajax()) {
+			abort(404);
+		}
+
+		if (Gate::denies('deleteCategory', Auth::user())) {
+			return redirect()->back()->with('error', 'no permission');
+			//return popup msg
+		}
+
 		$this->categories->destroyById($request->category_id);
-//		return redirect()->route('adminCategories')
-//			->with('info', 'The category was deleted successfully');
 	}
+
+
+	public function toogleCategoryActive(Request $request)
+	{
+		if (Gate::denies('toogleCategoryActive', $request->user())) {
+			abort(403);
+		}
+		
+		if ($request->ajax()) {
+			$this->categories->toogleCategoryActive($request);
+		} else {
+			abort(404);
+		}
+	}
+	
+	public function toogleCategorySeen(Request $request)
+	{
+		if (Gate::denies('toogleCategorySeen', $request->user())) {
+			abort(403);
+		}
+		
+		if ($request->ajax()) {
+			$this->categories->toogleCategorySeen($request);
+		} else {
+			abort(404);
+		}
+	}
+	
+	
 }
